// src/models/startUp/VMaccessRightRoleMapping.ts
/**
 *Created By: Md. Nazmus Salahin
 *Created Date: 10/25/2017
 *Modified By:
 *Modified date:
 *(C) CopyRight Nybsys ltd.
 */
import RoleModel from "./roleModel";

export default class VMaccessRightRoleMapping {

    accessRightRoleMappingID: number;
    accessRightID: number;
    status: number;
    rolesList: RoleModel[];
    constructor(){

    }
}